move ValidDuringRange model to its own file
